[household_expenses] Decimal instead of float

+++ household_expenses/forms.py
@@ -1,8 +1,8 @@
 from flask_wtf import FlaskForm
-from wtforms import BooleanField, FloatField, StringField
+from wtforms import BooleanField, DecimalField, StringField
 from wtforms.validators import DataRequired
 
 class ExpensesForm(FlaskForm):
     title = StringField('Nazwa', validators=[DataRequired()])
-    amount = FloatField('Kwota', validators=[DataRequired()])
+    amount = DecimalField('Kwota', validators=[DataRequired()])
     paid = BooleanField ('Zapłacone?')
